Moves all gRPC annotations to a new API package.
