+++ bom/pom.xml
@@ -102,6 +102,11 @@
             </dependency>
             <dependency>
                 <groupId>io.helidon.microprofile.grpc</groupId>
+                <artifactId>helidon-microprofile-grpc-api</artifactId>
+                <version>${helidon.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.helidon.microprofile.grpc</groupId>
                 <artifactId>helidon-microprofile-grpc-core</artifactId>
                 <version>${helidon.version}</version>
             </dependency>
