+++ microprofile/grpc/api/src/main/java/io/helidon/microprofile/grpc/api/GrpcInterceptor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.microprofile.grpc.core;
+package io.helidon.microprofile.grpc.api;
 
 import java.io.Serial;
 import java.lang.annotation.Documented;
