+++ microprofile/grpc/core/src/main/java/module-info.java
@@ -21,6 +21,7 @@ module io.helidon.microprofile.grpc.core {
 
     requires io.helidon.common;
     requires io.helidon.microprofile.config;
+    requires io.helidon.microprofile.grpc.api;
 
     requires transitive io.helidon.grpc.core;
     requires transitive jakarta.cdi;
