+++ microprofile/grpc/core/src/test/java/io/helidon/microprofile/grpc/core/AnnotatedMethodTest.java
@@ -19,6 +19,9 @@ package io.helidon.microprofile.grpc.core;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 
+import io.helidon.microprofile.grpc.api.Grpc;
+import io.helidon.microprofile.grpc.api.GrpcMethod;
+
 import io.grpc.MethodDescriptor;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
