+++ microprofile/grpc/core/src/test/java/io/helidon/microprofile/grpc/core/ModelHelperTest.java
@@ -19,6 +19,8 @@ package io.helidon.microprofile.grpc.core;
 import java.util.AbstractMap;
 
 import io.helidon.grpc.core.MarshallerSupplier;
+import io.helidon.microprofile.grpc.api.Grpc;
+import io.helidon.microprofile.grpc.api.GrpcMarshaller;
 
 import org.junit.jupiter.api.Test;
