+++ microprofile/grpc/pom.xml
@@ -32,6 +32,7 @@
     <packaging>pom</packaging>
 
     <modules>
+        <module>api</module>
         <module>core</module>
         <module>server</module>
     </modules>
