+++ microprofile/grpc/server/src/test/java/io/helidon/microprofile/grpc/server/BaseServiceTest.java
@@ -21,6 +21,7 @@ import io.helidon.common.tls.Tls;
 import io.helidon.microprofile.testing.junit5.AddExtension;
 import io.helidon.microprofile.testing.junit5.HelidonTest;
 import io.helidon.webclient.grpc.GrpcClient;
+
 import jakarta.ws.rs.client.WebTarget;
 
 @HelidonTest
