+++ microprofile/grpc/server/src/test/java/io/helidon/microprofile/grpc/server/EchoServiceTest.java
@@ -21,21 +21,22 @@ import java.lang.annotation.Target;
 import java.util.HashSet;
 import java.util.Set;
 
+import io.helidon.grpc.core.ContextKeys;
+import io.helidon.microprofile.grpc.api.Grpc;
+import io.helidon.microprofile.grpc.api.GrpcInterceptor;
+import io.helidon.microprofile.grpc.api.GrpcInterceptorBinding;
+import io.helidon.microprofile.grpc.api.GrpcInterceptors;
+import io.helidon.microprofile.grpc.api.Unary;
+import io.helidon.microprofile.grpc.server.test.Echo;
+import io.helidon.microprofile.grpc.server.test.EchoServiceGrpc;
+import io.helidon.tracing.Tracer;
+
 import io.grpc.Context;
 import io.grpc.Metadata;
 import io.grpc.ServerCall;
 import io.grpc.ServerCallHandler;
 import io.grpc.ServerInterceptor;
 import io.grpc.stub.StreamObserver;
-import io.helidon.grpc.core.ContextKeys;
-import io.helidon.microprofile.grpc.core.Grpc;
-import io.helidon.microprofile.grpc.core.GrpcInterceptor;
-import io.helidon.microprofile.grpc.core.GrpcInterceptorBinding;
-import io.helidon.microprofile.grpc.core.GrpcInterceptors;
-import io.helidon.microprofile.grpc.core.Unary;
-import io.helidon.microprofile.grpc.server.test.Echo;
-import io.helidon.microprofile.grpc.server.test.EchoServiceGrpc;
-import io.helidon.tracing.Tracer;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.client.WebTarget;
