+++ microprofile/grpc/server/src/test/java/io/helidon/microprofile/grpc/server/GrpcServiceBuilderTest.java
@@ -20,15 +20,16 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import io.helidon.microprofile.grpc.api.Grpc;
+import io.helidon.microprofile.grpc.api.GrpcMarshaller;
+import io.helidon.microprofile.grpc.api.GrpcMethod;
+import io.helidon.webserver.grpc.GrpcMethodDescriptor;
+import io.helidon.webserver.grpc.GrpcServiceDescriptor;
+
 import io.grpc.Metadata;
 import io.grpc.ServerCall;
 import io.grpc.ServerCallHandler;
 import io.grpc.stub.StreamObserver;
-import io.helidon.microprofile.grpc.core.Grpc;
-import io.helidon.microprofile.grpc.core.GrpcMarshaller;
-import io.helidon.microprofile.grpc.core.GrpcMethod;
-import io.helidon.webserver.grpc.GrpcMethodDescriptor;
-import io.helidon.webserver.grpc.GrpcServiceDescriptor;
 import jakarta.enterprise.inject.Instance;
 import jakarta.enterprise.inject.spi.BeanManager;
 import jakarta.inject.Singleton;
